styles for movie tile moved to the seperate component in common folder
