+++ src/features/MoviesList/index.js
@@ -1,148 +1,13 @@
 import { Container } from "../../common/Container/styled";
-import { Content, List, MainHeader, Poster, Tag, Tags, Title, Tile, Year, Star, Rating, Rate, Votes } from "./styled";
-import poster from "../../common/img/poster.png";
+import { List, MainHeader } from "./styled";
+import { MovieTile } from "../../common/MovieTile";
 
 export const MovieList = () => {
     return (
         <Container>
             <MainHeader>Popular Movies</MainHeader>
             <List>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
-                <Tile>
-                    <Poster src={poster} alt="movie poster" />
-                    <Content>
-                        <Title>Mulan</Title>
-                        <Year>2020</Year>
-                        <Tags>
-                            <Tag>Action</Tag>
-                            <Tag>Adventure</Tag>
-                            <Tag>Drama</Tag>
-                        </Tags>
-                        <Rating>
-                            <Star />
-                            <Rate>7,8</Rate>
-                            <Votes>35 votes</Votes>
-                        </Rating>
-                    </Content>
-                </Tile>
+                <MovieTile />
             </List>
         </Container>
     )
